改进数据导入，订单库存显示功能

+++ GODInventory.MyLinq/v_pendingorder.cs
@@ -189,7 +189,7 @@ namespace GODInventory.MyLinq
 
         public int 社内伝番処理 { get; set; }
        
-        public decimal 運賃 { get; set; }
+        public int 運賃 { get; set; }
         
         // 打印报表时使用。
         public int SubOrderCount { get; set; }
